Make main code consistent
This change updates each entry point of our microservices to the following in a more consistent way.

- Use cobra for argument parsing and help.
- Accept the configuration file using `--config-file` as a commandline option.
- Use a consistent pattern for choosing the ports used for various purposes.

The port changes only affect local development. In Kuberentes each pod can choose its own listening port without conflict.
