+++ deploy/Chart/templates/ucp/deployment.yaml
@@ -33,9 +33,9 @@ spec:
       containers:
       - name: ucp
         image: "{{ .Values.ucp.image }}:{{ .Values.ucp.tag | default $appversion }}"
+        args:
+        - --config-file=/etc/config/ucp-config.yaml
         env:
-        - name: UCP_CONFIG
-          value: /etc/config/ucp-config.yaml
         - name: BASE_PATH
           value: '/apis/api.ucp.dev/v1alpha3' # listen for APIService URLs
         - name: TLS_CERT_DIR
